+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/Decomposer/Output.cs
@@ -2,7 +2,7 @@ using NBitcoin;
 using System.Diagnostics.CodeAnalysis;
 using WalletWasabi.Extensions;
 
-namespace WalletWasabi.WabiSabi.Client.CoinJoin.Client.Decomp;
+namespace WalletWasabi.WabiSabi.Client.CoinJoin.Client.Decomposer;
 
 public record Output
 {
