+++ WalletWasabi/WabiSabi/Client/CoinJoin/Client/UtxoSelectionParameters.cs
@@ -6,7 +6,7 @@ using WabiSabi.Crypto.Randomness;
 using WalletWasabi.Extensions;
 using WalletWasabi.WabiSabi.Backend.Rounds;
 using WalletWasabi.WabiSabi.Models;
-using WalletWasabi.WabiSabi.Client.CoinJoin.Client.Decomp;
+using WalletWasabi.WabiSabi.Client.CoinJoin.Client.Decomposer;
 
 namespace WalletWasabi.WabiSabi.Client;
